Ensure module names are correctly displayed.
Case CPANEL-6109: There are two ways for an assessor module to provide
advice: call a method on the security advisor object, or call a method
on $self that wraps the method on the security advisor object.  If the
latter case occurred, the lookup for the function which called us would
always produce something in the Cpanel::Security::Advisor::Assessors
package.

The right way to solve this would be to inspect the class of the $self
object whose method called us, but that isn't possible.  Instead, look
for a specific module pattern either 1 or 3 steps back (to ignore the 2
levels of helper functions) and take that as the module name.