Check CSF Ports
Previously this module only checked for iptables rules that would deny
the mysql port, however by default csf denies all ports, and then allows
only those specified in the configuration, so the rules it's looking for
would not be present. In addition to the check against iptables, I added
a check to see if the csf.conf file exists, and whether or not the mysql
port is open. Also, I modified the suggestions to recommend either
changing the bind address OR closing the port.